Création d'un contrôleur pour uniformiser les messages d'erreurs

+++ backend/oc-chatop-backend/src/main/java/fr/gharrowbm/occhatopbackend/models/BaseExceptionResponse.java
@@ -4,6 +4,6 @@ public record BaseExceptionResponse(
         String message,
         String timestamp,
         String path,
-        Long statusCode
+        Integer statusCode
 ) {
 }
